LayoutTests/accessibility/mac/aria-pressed-button-attributes-expected.txt

This test ensures a role=button element with aria-pressed has the right attribute values, even after dynamic page changes.

#button1: AXRole: AXCheckBox
#button1: AXSubrole: AXToggle
#button1: AXValue: 1
#button1: AXARIAPressedIsPresent: true

#button2: AXRole: AXCheckBox
#button2: AXSubrole: AXToggle
#button2: AXValue: 0
#button2: AXARIAPressedIsPresent: true

#button-without-initial-press: AXRole: AXButton
#button-without-initial-press: AXSubrole:
#button-without-initial-press: AXValue:
#button-without-initial-press: AXARIAPressedIsPresent: false

Toggling aria-pressed state on both buttons.

#button1: AXRole: AXCheckBox
#button1: AXSubrole: AXToggle
#button1: AXValue: 0
#button1: AXARIAPressedIsPresent: true

#button2: AXRole: AXCheckBox
#button2: AXSubrole: AXToggle
#button2: AXValue: 1
#button2: AXARIAPressedIsPresent: true

#button-without-initial-press: AXRole: AXButton
#button-without-initial-press: AXSubrole:
#button-without-initial-press: AXValue:
#button-without-initial-press: AXARIAPressedIsPresent: true


PASS successfullyParsed is true

TEST COMPLETE
Foo
Bar
Baz
